refactor(activerecord/encryption): break circular import, fix empty-key check, drop unknown cast
- Extract key provider cache to default-key-provider-cache.ts; both
  scheme.ts and encryptor.ts import from it — no scheme↔encryptor cycle
- buildEncryptedMessage: 'if (!key)' → 'if (key == null)' so empty-string
  keys pass through and get rejected by the cipher (consistent with encrypt/decrypt)
- defaultKeyProvider return type is now KeyProviderLike; DerivedSecretKeyProvider
  is structurally compatible so no cast needed
