+++ packages/activerecord/src/encryption/encryptor.ts
@@ -8,7 +8,7 @@ import { Cipher } from "./cipher/aes256-gcm.js";
 import { Message } from "./message.js";
 import { MessageSerializer } from "./message-serializer.js";
 import { Configurable } from "./configurable.js";
-import { getOrCreateDefaultKeyProvider } from "./scheme.js";
+import { getOrCreateDefaultKeyProvider } from "./default-key-provider-cache.js";
 import { ConfigError, DecryptionError, ForbiddenClass } from "./errors.js";
 import type { Compressor } from "./config.js";
 import { defaultCompressor } from "./config.js";
@@ -131,14 +131,9 @@ export class Encryptor {
     if (ctxKp) return ctxKp;
     const { primaryKey, keyDerivationSalt, hashDigestClass } = Configurable.config;
     if (!primaryKey) return undefined;
-    // Reuse scheme.ts's module-level cache (keyed by primaryKey+salt+digest) so
-    // PBKDF2 runs once per config tuple and invalidation piggybacks on the
-    // single onConfigure hook already registered there.
-    return getOrCreateDefaultKeyProvider(
-      primaryKey,
-      keyDerivationSalt,
-      hashDigestClass,
-    ) as unknown as KeyProviderLike;
+    // Module-level cache keyed by (primaryKey, salt, digest); invalidated by
+    // the single onConfigure hook registered in default-key-provider-cache.ts.
+    return getOrCreateDefaultKeyProvider(primaryKey, keyDerivationSalt, hashDigestClass);
   }
 
   /** @internal */
@@ -178,7 +173,7 @@ export class Encryptor {
   ): Message {
     const encKeyObj = keyProvider.encryptionKey();
     const key = encKeyObj.secret;
-    if (!key) throw new ConfigError("No encryption key provided");
+    if (key == null) throw new ConfigError("No encryption key provided");
 
     const [cipherInput, compressed] = this.compressIfWorthIt(clearText);
     const { payload, iv, authTag } = this.cipher().encrypt(cipherInput, key, cipherOptions);
