+++ packages/activerecord/src/encryption/scheme.ts
@@ -12,50 +12,14 @@ import { withEncryptionContext } from "./context.js";
 import { DerivedSecretKeyProvider } from "./derived-secret-key-provider.js";
 import { DeterministicKeyProvider } from "./deterministic-key-provider.js";
 import { Key } from "./key.js";
-
-// Module-level single-entry cache for the default key provider. Shared across
-// all Scheme instances so PBKDF2 runs once per (primaryKey, salt, digest) tuple.
-// Uses stable string serialisation for primaryKey so array instances compare by
-// value rather than reference, avoiding spurious re-derivations.
-let _defaultKeyProviderEntry: DerivedSecretKeyProvider | undefined;
-let _defaultKeyProviderSig: string | undefined;
-
-function stableKeySignature(
-  primaryKey: string | string[],
-  keyDerivationSalt: string | undefined,
-  hashDigestClass: string,
-): string {
-  // JSON.stringify preserves array order (key rotation order is semantically
-  // meaningful) and is unambiguous — no comma-collision risk from key strings.
-  // Use null for undefined so missing salt is distinct from empty-string salt,
-  // ensuring cache invalidation when keyDerivationSalt is cleared.
-  // Normalize digest the same way KeyGenerator does (lowercase, no hyphens)
-  // so "SHA-256" and "sha256" resolve to the same cache entry.
-  const digest = hashDigestClass.toLowerCase().replace(/-/g, "");
-  return JSON.stringify([primaryKey, keyDerivationSalt ?? null, digest]);
-}
-
-export function getOrCreateDefaultKeyProvider(
-  primaryKey: string | string[],
-  keyDerivationSalt: string | undefined,
-  hashDigestClass: string,
-): DerivedSecretKeyProvider {
-  const sig = stableKeySignature(primaryKey, keyDerivationSalt, hashDigestClass);
-  if (!_defaultKeyProviderEntry || _defaultKeyProviderSig !== sig) {
-    _defaultKeyProviderEntry = new DerivedSecretKeyProvider(primaryKey);
-    _defaultKeyProviderSig = sig;
-  }
-  return _defaultKeyProviderEntry;
-}
-
-export function clearDefaultKeyProviderCache(): void {
-  _defaultKeyProviderEntry = undefined;
-  _defaultKeyProviderSig = undefined;
-}
-
-// Register eager cache invalidation so key material is released whenever
-// Configurable.configure() is called (key rotation, test teardown, etc.).
-Configurable.onConfigure(clearDefaultKeyProviderCache);
+import {
+  getOrCreateDefaultKeyProvider,
+  clearDefaultKeyProviderCache,
+} from "./default-key-provider-cache.js";
+export {
+  getOrCreateDefaultKeyProvider,
+  clearDefaultKeyProviderCache,
+} from "./default-key-provider-cache.js";
 
 export interface SchemeOptions {
   keyProvider?: unknown;
